feat: reference data download (#9)
* feat: initial work on routes for downloading ref data

* refactor: redirect_uri is handled by rewriting rules

* feat: sort roles, improve test

* refactor: tweak to test execution

* feat: add note to add auth

* fix: ensure response is returned for errors

* fix: typo in connection string

* chore: install deps for data conversion

* refactor: clarify module in use

* feat: add function for getting organisation list

* feat: extend download ref data feature

* fix: lint issues

* feat: cache getStandardisedOfficeLocationMap due to additional use

* fix: mock new server method

* feat: implement download for office locations

* feat: add blurbs explaining file download

* feat: correct name and add test

* feat: add tests for data ref routes

+++ server/routes/message-send.js
@@ -49,6 +49,8 @@ module.exports = [
     path,
     handler: async (request, h) => {
       const message = await verifyRequest(request)
+      if (message.isBoom) { return message }
+
       const users = await request.server.methods.db.getUsers()
 
       const phoneNumbersToSendTo = getPhoneNumbersToSendTo(users, message)
@@ -74,6 +76,8 @@ module.exports = [
     path,
     handler: async (request, h) => {
       const message = await verifyRequest(request)
+      if (message.isBoom) { return message }
+
       // Drop from cache to run a fresh query, getting the most uptodate info
       await request.server.methods.db.getUsers.cache.drop()
       const users = await request.server.methods.db.getUsers()
